move smu src in folder
