+++ src/smu/codec.rs
